Add web app link picker
Mirrors the TUI's ctrl+l fuzzy link picker so http/https URLs the agent
emitted in terminal output can be opened from the PWA. Pure extraction
moves to internal/links/ so the API handler and TUI share one
implementation. Esc is dropped from the overflow menu since it
duplicates the key bar.
