+++ README.md
@@ -305,11 +305,12 @@ Open `http://<your-machine>:7743/` in your phone browser. Enter the API token fr
 The dashboard provides:
 
 - **Task list** — Active and Archived tabs, status badges, free-text search box (case-insensitive substring against task name and project name; matching projects auto-expand so results aren't hidden by collapsed groups; `×` clears)
-- **Task detail** — Real xterm.js terminal with live SSE byte stream, plus an overflow (⋯) menu containing Esc / Toggle mode (Shift+Tab) / Resume / Files & diff / View prompt / Input history / Upload files / Rename / Fork / Archive / Delete and font size controls. The header auto-compacts when the soft keyboard is up. Live writes pause while you scroll into history (preserves iOS momentum-scroll); a green dot on the jump-to-input button shows new output is queued, tapping it flushes and returns to the live tail
+- **Task detail** — Real xterm.js terminal with live SSE byte stream, plus an overflow (⋯) menu containing Toggle mode (Shift+Tab) / Resume / Files & diff / View prompt / Input history / Open link / Upload files / Rename / Fork / Archive / Delete and font size controls (Esc lives in the key bar below). The header auto-compacts when the soft keyboard is up. Live writes pause while you scroll into history (preserves iOS momentum-scroll); a green dot on the jump-to-input button shows new output is queued, tapping it flushes and returns to the live tail
 - **Compose bar (touch only)** — A real native textarea sits below the terminal on phones/tablets. Type or dictate into it (iOS dictation, third-party keyboards, Wispr Flow) then tap **↑** or press **Enter** to send; **Shift+Enter** inserts a newline. xterm.js's hidden helper textarea is invisible to iOS text-injection — the compose bar is the on-screen field those tools need. Type `/` to open the skill autocomplete dropdown (Enter accepts a selection, Escape closes without inserting). Desktop's direct-into-xterm path is unchanged
 - **Key bar (touch only)** — Tap **⌨** in the compose bar to expose virtual keys iOS doesn't surface on its soft keyboard: **Esc**, **Tab**, **⇧Tab** (cycle Claude Code mode), **↵** (Enter), and **↑ ↓ ← →**. Buttons send the raw escape sequence to the agent (same path as the overflow Esc / Toggle mode items), so they navigate Claude's menus exactly like a hardware keypress. Visibility persists per device. Tapping a key keeps the soft keyboard up so you can type, navigate a menu, then send
 - **View prompt** — Tap **View prompt** in the overflow menu to see the original prompt the task was created with in a read-only modal (scrollable, monospace, wraps long lines). The modal exposes a **Copy** button next to **Close** that writes the prompt body to the clipboard
 - **Input history** — Tap **Input history** in the overflow menu to see the original prompt plus every line you've submitted from the live terminal (segmented on Enter, with timestamps). Captured client-side from each `/input` POST, persisted in `localStorage` per task so it survives reloads. ANSI/CSI sequences (arrow keys, Toggle mode, bracket-paste markers) and other control bytes are filtered out — only printable user text is recorded. **Clear** wipes the persisted history for the current task
+- **Open link** — Tap **Open link** in the overflow menu to surface every http/https URL the agent emitted in this task's terminal output. ANSI escape sequences, OSC 8 hyperlinks, and inline trailing punctuation are stripped before extraction so URLs surface cleanly. The modal has a substring filter (Enter opens the first match); tapping a row opens the URL in a new tab — universal-link-backed providers (GitHub, GitLab) are routed via `location.href` on iOS PWA installs so the system hands them off to the native app instead of leaving an `about:blank` stub. Mirrors the TUI's `ctrl+l` link picker
 - **File uploads** — Tap **Upload files** in the overflow menu to attach files mid-session (camera, photo library, or any file). Saved to `<worktree>/.context/`, paths are pasted into the agent's stdin (no Enter) so you can add a follow-up sentence before submitting. The New Task form has an inline drop zone with the same caps (10MB per file, 50MB total, 20 files) — paths are appended to the prompt automatically
 - **Files & diff** — Tap **Files & diff** in the overflow menu to open a full-screen view of every changed file in the task's worktree (uncommitted + committed-on-branch). Tap a file to load its diff with `+`/`-`/`@@` coloring; the file list auto-collapses behind a tappable bar so the diff fills nearly the whole screen. Long source lines soft-wrap by default — toggle the **Wrap** button in the bar to fall back to horizontal scroll. Closing the view returns to the live terminal at full size — no re-attach
 - **Create tasks** — Pick the agent (Claude Code or Codex, defaults to the configured global default), select a project (defaults to the project folder you have open in the task list), enter a prompt, start a new agent. Skill autocomplete (type `/`) suggests per-project and global skills
@@ -344,6 +345,7 @@ All endpoints require auth — either `Authorization: Bearer <token>` header or 
 | Method | Endpoint                 | Description                                                                                                                                                                                             |
 | ------ | ------------------------ | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
 | `GET`  | `/api/tasks/{id}/output` | Recent output (text). Optional `?bytes=`, `?clean=1`                                                                                                                                                    |
+| `GET`  | `/api/tasks/{id}/links`  | Extract http/https URLs from terminal output. Returns `{"links":[{"label","url"}]}`. Mirrors the TUI's `ctrl+l` link picker; powers the web app's "Open link" overflow item                            |
 | `POST` | `/api/tasks/{id}/input`  | Send raw bytes to PTY stdin                                                                                                                                                                             |
 | `POST` | `/api/tasks/{id}/upload` | Upload files mid-session. `multipart/form-data` with `files` parts; saved to `<worktree>/.context/<name>` (auto-suffixed on collision) and returns `{paths:[]}`. Same 10MB/50MB/20-file caps as create. |
 | `GET`  | `/api/tasks/{id}/stream` | SSE stream of live output (base64-encoded chunks)                                                                                                                                                       |
