+++ context/knowledge/gotchas/misc.md
@@ -124,7 +124,8 @@
 
 ## Link Extraction
 
-- **`osc8Re` must run BEFORE `ansiRe` in `stripANSI`.** OSC 8 hyperlinks embed URLs in escape sequences (`\x1b]8;;URL\x1b\\`). If `ansiRe` strips them first, the URLs are lost. The two-pass design in `links.go:stripANSI` is intentional.
-- **`tui/ansiRe` and `sanitize/ansiRe` must both handle ST-terminated OSC (`\x1b\\`).** Claude Code uses ST (not BEL) for OSC 8 hyperlinks. Missing ST support causes URLs to splice with display text.
+- **`osc8Re` must run BEFORE `ansiRe` in `stripANSI`.** OSC 8 hyperlinks embed URLs in escape sequences (`\x1b]8;;URL\x1b\\`). If `ansiRe` strips them first, the URLs are lost. The two-pass design in `internal/links/links.go:stripANSI` is intentional.
+- **`tui/widget.AnsiRe`, `links/ansiRe`, and `sanitize/ansiRe` must all handle ST-terminated OSC (`\x1b\\`).** Claude Code uses ST (not BEL) for OSC 8 hyperlinks. Missing ST support causes URLs to splice with display text. The three regexes are intentionally separate: widget strips for display, links preserves OSC 8 URLs for extraction, sanitize strips for clean text.
+- **Link extraction is shared between TUI and HTTP API.** `internal/links` exports `Link` + `Extract`; `internal/tui/links.go` adds the modal + `openURL` helper, and `internal/api/handlers.go:handleGetLinks` reuses the same extractor over the live ring buffer (or session log tail, capped at 1MB). Don't re-implement extraction in either client — both must produce identical results so users moving between TUI `ctrl+l` and the web "Open link" overflow item see the same set.
 - **SGR sequences must strip to empty; cursor movement must strip to space.** Raw PTY logs contain cursor movement codes (`\x1b[5C`, `\x1b[1B`) that position text on different screen locations. Stripping these to empty merges unrelated text into URLs. `stripANSI` uses `ReplaceAllStringFunc` to distinguish: CSI ending in `m` → empty (preserves mid-URL colors), all else → space (prevents merging).
 - **`bareLinkRe` must exclude `"`, backtick, `{`, `}`, `<`.** These are never valid in URLs per RFC 3986. Without exclusion, the regex matches through quoted/JSON URL data producing garbage entries in the link picker.
