+++ context/knowledge/gotchas/web-remote.md
@@ -117,7 +117,8 @@ Non-obvious invariants for the SPA + REST API + push notifications stack.
 - **The service worker must intercept `/share` and serve the cached `/` shell, not let it fall through** — without the intercept, every share creates a unique cache entry keyed by query string (waste), and offline shares fail because the network request never completes. The intercept ignores query string and returns the same shell.
 - **Capture share params into `sessionStorage` BEFORE auth, apply them AFTER `tryConnect()` succeeds** — the share IIFE runs at script load and stashes `argus-pending-share`; `applyPendingShare()` is called from the success branch of `tryConnect`. Putting the apply step earlier means it runs against an unmounted New Task tab, since `loadProjects()` hasn't populated the dropdown yet. Putting the capture step later means a slow auth flow loses the params (history.replaceState already cleaned the URL).
 - **`history.replaceState(null, '', '/')` in the capture IIFE strips `/share` from the address bar** — without it, a refresh re-fires the share with the same content and the user gets a duplicate prefilled prompt every reload. The cached `/` shell controls the URL going forward.
-- **`window.open(url, '_blank')` in iOS standalone PWA leaves a stub `about:blank` window the user has to dismiss with X.** A standalone PWA has no browser chrome to host a real new tab, so `_blank` opens an empty placeholder. Universal links still hand off to the native app (e.g. github.com → GitHub app), which is what makes it look like the open "worked" — but the stub stays behind. Workaround in the View PR handler: if `navigator.standalone` is true and the URL is on github.com, assign `window.location.href` instead — iOS routes to the GitHub app and the PWA stays alive in the background. Don't blanket-replace `window.open`; for non-universal-link URLs, `location.href` would navigate the PWA away.
+- **`window.open(url, '_blank')` in iOS standalone PWA leaves a stub `about:blank` window the user has to dismiss with X.** A standalone PWA has no browser chrome to host a real new tab, so `_blank` opens an empty placeholder. Universal links still hand off to the native app (e.g. github.com → GitHub app), which is what makes it look like the open "worked" — but the stub stays behind. The shared `openExternalURL(url)` helper centralizes the workaround: if `navigator.standalone` is true and the URL matches `universalLinkRe` (github.com / gitlab.com), assign `window.location.href` instead — iOS routes to the native app and the PWA stays alive in the background. Don't blanket-replace `window.open`; for non-universal-link URLs, `location.href` would navigate the PWA away. Both the View PR overflow item and the link picker funnel through this helper — extend `universalLinkRe` if other providers are added (jira, linear).
+- **Link picker re-fetches `/api/tasks/{id}/links` every open.** A long-running agent can emit new URLs after the modal was last opened; caching the list across opens would silently hide them. The endpoint reads from the live ring buffer when the session is running and falls back to the on-disk session log tail (capped at 1MB) otherwise — so finished tasks still surface their URLs without keeping the session alive. Filter rows render as `<button>` tap targets so iOS keyboard focus + Enter activates the same path as a tap, and labels/URLs use `textContent` (never `innerHTML`) because the strings come straight from agent stdout.
 - **iOS Safari does not implement `share_target`; the manifest entry only works on Android Chromium.** Settings → "iOS share sheet" exposes `${origin}/share?text=[Shortcut Input]` for the user to paste into an iOS Shortcut (Show in Share Sheet enabled, Text + URLs only) — the literal `[Shortcut Input]` is the placeholder Shortcuts replaces with its magic variable. The "Test" button in that section navigates the current PWA to `/share?text=...` (NOT `window.open`, which would leave an `about:blank` stub in standalone mode) so the capture IIFE runs end-to-end. If you later add additional share params, also widen the Shortcut URL builder in `shareShortcutURL()` and update the Playwright assertion.
 
 ## New Task form
