+++ context/knowledge/index.md
@@ -11,7 +11,7 @@ Non-obvious invariants and gotchas, split by topic. Read the relevant file when 
 | [gotchas/worktree.md](gotchas/worktree.md)         | Worktree creation ordering, transactional CreateAndStart, cleanup, path validation, stale ref pruning, missing-worktree pre-flight                                                                                                                                                                                                                                                              | 14      |
 | [gotchas/keybindings.md](gotchas/keybindings.md)   | Key routing, ctrl sequences, tcell modifier quirks, agent view navigation                                                                                                                                                                                                                                                                                                                       | 12      |
 | [gotchas/tasklist-ui.md](gotchas/tasklist-ui.md)   | Task list cursor, modals, focus guards, filter, archive, spinner, row rendering, form done-flag reset                                                                                                                                                                                                                                                                                           | 26      |
-| [gotchas/web-remote.md](gotchas/web-remote.md)     | SPA + REST API + service worker, EventSource auth, xterm.js, HTML escaping, virtual keys, detail-view layout, files-view overlay, prompt modal, input history, new task form defaults, file uploads, offline view, Web Push, per-device tokens, settings endpoints, share target + iOS Shortcut helper, test harness, idle task status, static-output replay, mobile compose bar, key bar, skill autocomplete, task-list search, diff wrap & files-bar | 132     |
+| [gotchas/web-remote.md](gotchas/web-remote.md)     | SPA + REST API + service worker, EventSource auth, xterm.js, HTML escaping, virtual keys, detail-view layout, files-view overlay, prompt modal, input history, link picker, new task form defaults, file uploads, offline view, Web Push, per-device tokens, settings endpoints, share target + iOS Shortcut helper, test harness, idle task status, static-output replay, mobile compose bar, key bar, skill autocomplete, task-list search, diff wrap & files-bar | 133     |
 | [gotchas/misc.md](gotchas/misc.md)                 | DB patterns, Go idioms, Codex, MCP, PRs, file explorer, quick-add, scheduled tasks, links, task auto-naming                                                                                                                                                                                                                                                                                     | 92      |
 
 ## Other Files
