+++ internal/api/static/sw.js
@@ -7,7 +7,7 @@
 //
 // Increment SW_VERSION on any shell change to bust the cache.
 
-const SW_VERSION = 'argus-shell-v18';
+const SW_VERSION = 'argus-shell-v19';
 const SHELL_ASSETS = [
   '/',
   '/manifest.webmanifest',
