+++ web-tests/tests/vkeys.spec.ts
@@ -64,23 +64,14 @@ test.describe('agent detail view chrome', () => {
     await expect(page.locator('.detail-header.compact .detail-subtitle')).toBeHidden();
   });
 
-  test('Esc menu item sends ESC byte and keeps menu open', async ({ page }) => {
-    await login(page);
+  // Esc lives in the key bar (see keybar.spec.ts) — the overflow menu Esc
+  // entry was removed to declutter the dropdown.
 
+  test('overflow menu has no Esc entry', async ({ page }) => {
+    await login(page);
     await page.locator('#btn-overflow').click();
     await expect(page.locator('#overflow-menu.open')).toBeVisible();
-    await expect(page.locator('#btn-esc')).toBeVisible();
-
-    const inputReq = page.waitForRequest(req =>
-      req.url().includes('/input') && req.method() === 'POST',
-      { timeout: 3000 }
-    );
-    await page.locator('#btn-esc').click();
-    const req = await inputReq;
-    expect(req.postData()).toBe('\x1b');
-
-    // Menu must stay open so the user can spam Esc to interrupt a runaway loop.
-    await expect(page.locator('#overflow-menu.open')).toBeVisible();
+    await expect(page.locator('#btn-esc')).toHaveCount(0);
   });
 
   test('Toggle mode menu item sends Shift+Tab (CSI Z) and keeps menu open', async ({ page }) => {
@@ -104,6 +95,65 @@ test.describe('agent detail view chrome', () => {
     await login(page);
     await page.locator('#btn-overflow').click();
     await expect(page.locator('#btn-stop')).toHaveCount(0);
+  });
+
+  test('link picker opens, filters, and clicks invoke openExternalURL', async ({ page }) => {
+    // Same SW-disable as the View PR test: the network-only /api/* SW handler
+    // bypasses page.route() in WebKit unless register() is no-op'd here.
+    await page.addInitScript(() => {
+      const sw = navigator.serviceWorker;
+      if (sw) {
+        Object.defineProperty(sw, 'register', {
+          configurable: true,
+          value: () => Promise.reject(new Error('disabled in test')),
+        });
+      }
+    });
+
+    // Stub the links endpoint with a known set so the test isn't sensitive
+    // to whatever the bash harness has emitted.
+    await page.route('**/api/tasks/*/links*', async (route) => {
+      await route.fulfill({
+        status: 200,
+        contentType: 'application/json',
+        body: JSON.stringify({
+          links: [
+            { label: 'Example Docs', url: 'https://example.com/docs' },
+            { label: 'https://github.com/foo/bar', url: 'https://github.com/foo/bar' },
+          ],
+        }),
+      });
+    });
+
+    await login(page);
+
+    // Capture window.open calls instead of letting the test browser navigate.
+    await page.evaluate(() => {
+      (window as any).__opens = [];
+      window.open = (url: any) => {
+        (window as any).__opens.push(String(url));
+        return null;
+      };
+    });
+
+    await page.locator('#btn-overflow').click();
+    await page.locator('#btn-links').click();
+
+    // Modal renders both rows.
+    await expect(page.locator('#links-modal.open')).toBeVisible();
+    const items = page.locator('#links-modal-body .links-item');
+    await expect(items).toHaveCount(2);
+
+    // Filter narrows the list.
+    await page.locator('#links-modal-filter').fill('docs');
+    await expect(items).toHaveCount(1);
+    await expect(items.first()).toContainText('Example Docs');
+
+    // Click → window.open invoked, modal closes.
+    await items.first().click();
+    await expect(page.locator('#links-modal.open')).toHaveCount(0);
+    const opens = await page.evaluate(() => (window as any).__opens);
+    expect(opens).toEqual(['https://example.com/docs']);
   });
 
   test('View PR appears without re-entering when pr_url becomes available', async ({ page }) => {
